+++ src/notebooks/pipes/Visualization/types/Gauge/index.tsx
@@ -1,6 +1,6 @@
 import icon from './icon'
 import properties from './properties'
-import options from './options'
+import GaugeOptions from './GaugeOptions'
 
 export default register => {
   register({
@@ -8,6 +8,6 @@ export default register => {
     name: 'Gauge',
     graphic: icon,
     initial: properties,
-    options: options,
+    options: GaugeOptions,
   })
 }
