+++ src/notebooks/pipes/Visualization/types/Heatmap/index.tsx
@@ -1,6 +1,6 @@
 import icon from './icon'
 import properties from './properties'
-import options from './options'
+import HeatmapOptions from './HeatmapOptions'
 
 export default register => {
   register({
@@ -8,6 +8,6 @@ export default register => {
     name: 'Heatmap',
     graphic: icon,
     initial: properties,
-    options: options,
+    options: HeatmapOptions,
   })
 }
