+++ src/notebooks/pipes/Visualization/types/Histogram/index.tsx
@@ -1,6 +1,6 @@
 import icon from './icon'
 import properties from './properties'
-import options from './options'
+import HistogramOptions from './HistogramOptions'
 
 export default register => {
   register({
@@ -8,6 +8,6 @@ export default register => {
     name: 'Histogram',
     graphic: icon,
     initial: properties,
-    options: options,
+    options: HistogramOptions,
   })
 }
