+++ src/notebooks/pipes/Visualization/types/Scatter/index.tsx
@@ -1,6 +1,6 @@
 import icon from './icon'
 import properties from './properties'
-import options from './options'
+import ScatterOptions from './ScatterOptions'
 
 export default register => {
   register({
@@ -8,6 +8,6 @@ export default register => {
     name: 'Scatter',
     graphic: icon,
     initial: properties,
-    options: options,
+    options: ScatterOptions,
   })
 }
