+++ src/notebooks/pipes/Visualization/types/Table/TableOptions.tsx
@@ -30,8 +30,7 @@ import {
 } from 'src/shared/constants/thresholds'
 import DraggableColumn from 'src/shared/components/draggable_column/DraggableColumn'
 import {resolveTimeFormat} from 'src/dashboards/utils/tableGraph'
-import {VisOptionProps} from '../../'
-import {TableViewProperties, FieldOption} from 'src/types'
+import {TableViewProperties, FieldOption, VisOptionProps} from 'src/types'
 import {MIN_DECIMAL_PLACES, MAX_DECIMAL_PLACES} from 'src/dashboards/constants'
 import {convertUserInputToNumOrNaN} from 'src/shared/utils/convertUserInput'
 import ThresholdsSettings from 'src/shared/components/ThresholdsSettings'
